Move Modal component to ui folder, update all neccessary paths in other components
