+++ app/src/components/ModalBoard/ModalBoard.tsx
@@ -1,7 +1,7 @@
 import { useState, useContext, useRef, FormEvent, ChangeEvent } from "react";
 import BoardsContentContext from "@context/BoardsContent/BoardsContent";
 import BoardModalContext from "@context/BoardModal/BoardModal";
-import Modal from "@components/Modal/Modal";
+import Modal from "@components/ui/Modal/Modal";
 import Input from "@components/ui/Input/Input";
 import Button from "@components/ui/Button/Button/Button";
 import styles from "./ModalBoard.module.scss";
